pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "konfuse"
version = "0.1.0"
description = "Merge any kubeconfig in one command. Rename on import. Never lose your existing config."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
dependencies = ["PyYAML>=5.4"]

[project.scripts]
konfuse = "konfuse:main"

[project.urls]
Homepage = "https://github.com/chameerar/konfuse"
Source = "https://github.com/chameerar/konfuse"
"Bug Tracker" = "https://github.com/chameerar/konfuse/issues"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.4"]

[tool.ruff]
target-version = "py38"
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
